chore: refactor before making tree view
