+++ src/templates/workspace/snippets/treeIndent.ts
@@ -1,0 +1,16 @@
+export const treeIndent = (depth: number) => {
+  if (depth === 0) {
+    return '';
+  }
+
+  const indents = [...Array(depth).keys()]
+    .map(() => {
+      return `
+        <div class="list_branch-indent"></div>
+      `;
+    })
+    .join('')
+    .trim();
+
+  return `<div class="list_branch-indent-box">${indents}</div>`;
+};
